requirements.txt
# At first, please configure torch as follows:
# pip install torch==2.6.0 torchvision==0.21.0 torchaudio==2.6.0 --index-url https://download.pytorch.org/whl/cu124
# or (depending on your cuda version)
# pip install torch==2.7.0 torchvision==0.22.0 torchaudio==2.7.0 --index-url https://download.pytorch.org/whl/cu128
transformers==4.46.0
wandb
webdataset
omegaconf
pyarrow
pandas
jaxtyping
typeguard
accelerate
deepspeed==0.15.4
lightning
datasets
image-reward
torchmetrics==1.6.2
math_verify 
peft
einops
huggingface_hub
tiktoken
# after installing above, please install trl from source, as follows:
# pip install git+https://github.com/huggingface/trl.git@0f88c179e30b3439467942a08c3190f624d5c423
# export HF_ENDPOINT=https://hf-mirror.com
# at first run, remember to use wandb login, or the choices *shown at the start of training* may be neglected!